Was not able to get to the full solution myself. Got help from chatgpt. Not entirely sure why my code doesnt work.. I also got it to spit out code that looked a lot different that what I ended up saving to study but not sure its a correct solution. Will be studying we're our logic differs
